Add get model methods to each adapter
